User team can use our comps! Fixed follow button.
User team can now use our components on the profile page. Made a follow and unfollow button work properly.
If you are on the same page as your logged in account, the FollowButton will turn into a Edit Profile button.
I made a testPage for people to try. The route is  /followCompsTestPage/id